requirements.txt
# Core dependencies
pyyaml
numpy
scipy
torch
torchvision
pillow
matplotlib

# Optional: for logging (standard library, but listed for clarity)
# logging

# If you use Jupyter notebooks, you may want:
# jupyter
